# Implementation/helloworldIvo/src/main/resources/public/swagger.yaml
...
swagger: '2.0'
info:
  description: 'Documentation for an API'
  version: '1.0'
  title: 'API documentation'
  contact:
    name: 'Ivo Gurtner and Sabrina Westhauser'
    email: 'ivo.gurtner@noseryoung.ch'
host: 'localhost:8080'
basePath: /
...
paths:
...
  /roles:
    get:
      tags:
        - RoleController
      summary: 'Returns all roles'
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/RoleDTO'
        403:
          description: Forbidden
      deprecated: false
    post:
      tags:
        - RoleController
      summary: 'Creates a new role'
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          required: true
          description: "The new roles data"
          schema:
            $ref: '#/definitions/RoleDTO'
      responses:
        201:
          description: Created
          schema:
            $ref: '#/definitions/RoleDTO'
        400:
          description: Bad Request
        403:
          description: Forbidden
      deprecated: false
  /roles/{id}:
    get:
      tags:
        - RoleController
      summary: 'Returns a single role with the provided ID'
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 'ID of the requested entity'
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/RoleDTO'
        403:
          description: Forbidden
        404:
          description: Not Found
      deprecated: false
    put:
      tags:
        - RoleController
      summary: 'Updates the role with the provided ID'
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 'ID of the requested entity'
          required: true
          type: string
        - in: body
          name: role
          description: "The updated roles data"
          required: true
          schema:
            $ref: '#/definitions/RoleDTO'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/RoleDTO'
        400:
          description: Bad Request
        403:
          description: Forbidden
        404:
          description: Not Found
      deprecated: false
    delete:
      tags:
        - RoleController
      summary: 'Deletes a role with the provided ID'
      produces:
        - none
      parameters:
        - name: id
          in: path
          description: 'ID of the requested entity'
          required: true
          type: string
      responses:
        204:
          description: No Content
        403:
          description: Forbidden
        404:
          description: Not Found
      deprecated: false

  # USER
  /users:
    get:
      tags:
        - UserController
      summary: 'Returns all users'
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/UserDTO'
        403:
          description: Forbidden
      deprecated: false
    post:
      tags:
        - UserController
      summary: 'Creates a new user'
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          required: true
          description: "The new users data"
          schema:
            $ref: '#/definitions/UserDTO'
      responses:
        201:
          description: Created
          schema:
            $ref: '#/definitions/UserDTO'
        400:
          description: Bad Request
        403:
          description: Forbidden
      deprecated: false
  /users/{id}:
    get:
      tags:
        - UserController
      summary: 'Returns a single user with the provided ID'
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 'ID of the requested entity'
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/UserDTO'
        403:
          description: Forbidden
        404:
          description: Not Found
      deprecated: false
    put:
      tags:
        - UserController
      summary: 'Updates the user with the provided ID'
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 'ID of the requested entity'
          required: true
          type: string
        - in: body
          name: user
          description: "The updated users data"
          required: true
          schema:
            $ref: '#/definitions/UserDTO'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/UserDTO'
        400:
          description: Bad Request
        403:
          description: Forbidden
        404:
          description: Not Found
      deprecated: false
    delete:
      tags:
        - UserController
      summary: 'Deletes a user with the provided ID'
      produces:
        - none
      parameters:
        - name: id
          in: path
          description: 'ID of the requested entity'
          required: true
          type: string
      responses:
        204:
          description: No Content
        403:
          description: Forbidden
        404:
          description: Not Found
      deprecated: false
  /users/salary/{min}/{max}:
    get:
      tags:
        - UserController
      summary: 'Returns a list of users with a given range of their salary'
      produces:
        - application/json
      parameters:
        - name: min
          in: path
          description: 'Min. salary'
          required: true
          type: integer
        - name: max
          in: path
          description: 'Max. salary'
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/UserDTO'
        403:
          description: Forbidden
        404:
          description: Not Found
      deprecated: false
  # ARTICLE
...
definitions:
...
  AuthorityDTO:
    # What is data type this object?
    type: object
    # What parts of it are required?
    required:
      - name
    # The effective properties of the object
    properties:
      id:
        type: string
        example: 'id'
      name:
        type: string
        example: 'ADMIN'
    # How the object is called
    title: AuthorityDTO

  # The DTOs from here on out are built in a similar way

  # ROLE
  RoleDTO:
    type: object
    required:
      - name
      - authorities
    properties:
      id:
        type: string
        example: 'id'
      name:
        type: string
        example: 'Admin'
      authorities:
        type: array
        items:
          $ref: '#/definitions/AuthorityDTO'
    title: RoleDTO

  # USER
  UserDTO:
    type: object
    required:
      - email
      - password
      - enabled
    properties:
      id:
        type: string
        example: 'id'
      email:
        type: string
        example: 'example@mail.com'
      password:
        type: string
        example: 'youshouldNotPutAPasswordHere'
      enabled:
        type: boolean
        example: true
      roles:
        type: array
        items:
          $ref: '#/definitions/RoleDTO'
    title: UserDTO

  # ARTICLE
...